Memory optimizations: Query and internal archetype list (#230)
* reduce size of Query to 80 bytes
* use uint32 for pool capacity increment

+++ ecs/archetypes_test.go
@@ -18,7 +18,7 @@ func TestArchetypePointers(t *testing.T) {
 	pt.Add(&a2)
 	pt.Add(&a3)
 
-	assert.Equal(t, 3, pt.Len())
+	assert.Equal(t, int32(3), pt.Len())
 
 	for i := 0; i < 45; i++ {
 		pt.Add(&a3)
